redecorating the code folders
