+++ cors-server/routes/register.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import * as https from "https";
+import {encryptPassword, validateUser} from "./authentication";
 
 module.exports = function (server, db){
 
@@ -15,16 +15,18 @@ module.exports = function (server, db){
         })
     })
 
-    server.post('/data/register', (request: express.Request, response: express.Response) => {
+    server.post('/data/register', async (request: express.Request, response: express.Response) => {
         const {username, email, password} = request.body;
-        const statment = 'INSERT INTO user (username, email, password, online) VALUES(?, ?, ?, ?)';
+        const encryptedPassword = await encryptPassword(password);
+        const query = 'INSERT INTO user (username, email, password, online) VALUES(?, ?, ?, ?)'
+
         try {
-            db.all(statment, [username, email, password, true], (err, rows) => {
-                    if (!err) {
-                        response.json({userCreated: true})
-                    } else {
-                        response.status(400).json({error: "username or email already in use", userCreated: false});
-                    }
+            db.all(query, [username, email, encryptedPassword, true], (err, rows) => {
+                if (!err) {
+                    response.json({userCreated: true})
+                } else {
+                    response.status(400).json({error: "username or email already in use", userCreated: false});
+                }
             });
         } catch (e) {
             console.log(e)
